Add files via upload
A retail dataset consists of 3 files; Features.csv, sales.csv and stores.csv files. As the names suggest, the files consist of different attributes in the features file such as temperature, fule price, CPI, unemployment rate, Markdown prices 1,2,3,4 and 5. The sales file consist of weekly sales figures in addition to store data.

An analysis of each attribute is done and its influence on the dependent variable, weekly-sales is studied. A multiple linear regression model is created and the weekly sales for the next year are predicted.